gcp-cf-src/requirements.txt
functions-framework==3.*
python-dotenv
vertexai
grpcio==1.64.0
grpcio-status==1.64.0
boto3
slack_sdk
dataclasses
google-cloud-secret-manager
jsonpickle